// src/app/modules/match/_interfaces/match-event.interface.ts
export interface MatchEvent {

  assignedCategoryId?: string;
  assignedCategoryTitle?: string;

  playMinute?: number;

  assignedPlayerOneId?: string;
  assignedPlayerOneName?: string;

  assignedPlayerTwoId?: string;
  assignedPlayerTwoName?: string;

  id?: string;
  title?: string;
  description?: string;
}